+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `asWrapper` function to cast native refs to wrapper refs (#30, #31)
 - `DataValue::getValuePtr` method
 - `Server::getNamespaceArray` method
+- `Server::runIterate` method, e.g. to run server in existing event loop
 
 ### Changed